Adding "printed" label
At my university we create scripts for the students, which are usually also made available in printed form for the major versions. I've been using the package for a long time and now came up with the idea that, like for "YANKED", you can also give a hint as to whether that version is also available in printed form. I would be happy if this function as shown here, or in another form, is implemented in the next version of "changelog".

Would it also be possible to create a boolean trigger for the version number, which sets it either in \textbf bold or in \texttt teletype?