Disable `\label` if `label=` in `changelog`
If the `label` option for the `changelog` environment is empty,
no `label` will be generated for the changelog section.

The documentation has been clarified to explain that if the
`section` option is set to `false`, no label will be generated, as
labels are meaningless without being attached to a section.

Closes #25.